Adding Exporting Libraries and Functionality
Issue with writing all activities to each sheet which is unintended and needs to be addressed.

Added ability to export/save to selected directory, overwrite if file already exists, cancel the save, as well as error handling if file is opened and thus cannot overwrite it and alerts the user.